i HATE HAVING TO PUBLISH THE BRANCH, SORRY YOU DIDN'T SEE THIS I WORKED THORUGH THE CHANGES.
That said this part is just a small snipet of what we need to do so you have some idea.

Basically for each production there's 3 things, first we have the parser.yy where we put the production.

Then the node definition in ast.hpp

Then the unparsing definition in unparse.cpp